requirements.txt
alembic==1.13.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
flake8==7.0.0
Flask==3.0.2
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.0
pytest-env==0.6.2
python-dateutil==2.9.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.12.0
Werkzeug==3.0.2
WTForms==3.1.2
zipp==3.18.1